+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [5.12.0] - 2025-07-17
+- added severity, service and resource type to operational rules report
+- changed operational overview report to new format
 
 
 ## [5.11.0] - 2025-06-13